predict.py: Ensure new sample is reshaped
New sample must have batch size of 1.
To do this, reshape to add axis=0 with dimensionality of 1

Otherwise, the NN will output a 2D tensor of shape (sequence_length, n_classes), and make a prediction for every word/token in the sequence. This isn't what we want - we want one prediction for the whole sequence (i.e. shape=(1, n_classes))